refactoring, fixes #80
