+++ app/src/main/java/com/ivoberger/enq/ui/fragments/SuggestionsFragment.kt
@@ -4,14 +4,14 @@ import android.os.Bundle
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import com.ivoberger.enq.persistence.AppSettings
-import com.ivoberger.enq.ui.fragments.parents.TabbedResultsFragment
+import com.ivoberger.enq.ui.fragments.base.TabbedResultsFragment
+import com.ivoberger.enq.utils.retryOnError
 import com.ivoberger.jmusicbot.JMusicBot
 import com.ivoberger.jmusicbot.model.MusicBotPlugin
 import kotlinx.android.synthetic.main.fragment_results.*
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.async
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
 import splitties.experimental.ExperimentalSplittiesApi
 import splitties.lifecycle.coroutines.PotentialFutureAndroidXLifecycleKtxApi
 import splitties.lifecycle.coroutines.lifecycleScope
@@ -22,7 +22,7 @@ class SuggestionsFragment : TabbedResultsFragment() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        mProviderPlugins = lifecycleScope.async(Dispatchers.IO) { JMusicBot.getSuggesters() }
+        mProviderPlugins = lifecycleScope.async { retryOnError { JMusicBot.getSuggesters() } }
         lifecycleScope.launch(Dispatchers.IO) {
             mProviderPlugins.await() ?: return@launch
             AppSettings.lastSuggester?.also {
@@ -32,13 +32,12 @@ class SuggestionsFragment : TabbedResultsFragment() {
     }
 
     override fun initializeTabs() {
-        lifecycleScope.launch(Dispatchers.IO) {
+        lifecycleScope.launch {
             val provider = JMusicBot.getSuggesters()
-            mFragmentPagerAdapter =
-                async {
-                    SuggestionsFragmentPager(childFragmentManager, provider)
-                }
-            withContext(Dispatchers.Main) { view_pager.adapter = mFragmentPagerAdapter.await() }
+            mFragmentPagerAdapter = async(Dispatchers.Default) {
+                SuggestionsFragmentPager(childFragmentManager, provider)
+            }
+            view_pager.adapter = mFragmentPagerAdapter.await()
         }
     }
